# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"UserRole": {
				"enum": [
					"admin",
					"standard"
				],
				"type": "string"
			},
			"User": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/UserRole"
					},
					"token": {
						"type": "string"
					},
					"bookings": {
						"items": {
							"$ref": "#/components/schemas/Booking"
						},
						"type": "array"
					},
					"createdDate": {
						"type": "string",
						"format": "date-time"
					},
					"updatedDate": {
						"type": "string",
						"format": "date-time"
					},
					"deletedDate": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"firstName",
					"lastName",
					"email",
					"role",
					"token",
					"bookings",
					"createdDate",
					"updatedDate",
					"deletedDate"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Booking": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"parkingSpot": {
						"$ref": "#/components/schemas/ParkingSpot"
					},
					"parkingSpotId": {
						"type": "number",
						"format": "double"
					},
					"startDate": {
						"type": "string",
						"format": "date-time"
					},
					"endDate": {
						"type": "string",
						"format": "date-time"
					},
					"createdDate": {
						"type": "string",
						"format": "date-time"
					},
					"updatedDate": {
						"type": "string",
						"format": "date-time"
					},
					"deletedDate": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"user",
					"userId",
					"parkingSpot",
					"parkingSpotId",
					"startDate",
					"endDate",
					"createdDate",
					"updatedDate",
					"deletedDate"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ParkingSpot": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"bookings": {
						"items": {
							"$ref": "#/components/schemas/Booking"
						},
						"type": "array"
					},
					"createdDate": {
						"type": "string",
						"format": "date-time"
					},
					"updatedDate": {
						"type": "string",
						"format": "date-time"
					},
					"deletedDate": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"bookings",
					"createdDate",
					"updatedDate",
					"deletedDate"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BookingPayload": {
				"properties": {
					"endDate": {
						"type": "string"
					},
					"startDate": {
						"type": "string"
					},
					"parkingSpotId": {
						"type": "number",
						"format": "double"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"endDate",
					"startDate",
					"parkingSpotId"
				],
				"type": "object"
			},
			"Partial_BookingPayload_": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"parkingSpotId": {
						"type": "number",
						"format": "double"
					},
					"startDate": {
						"type": "string"
					},
					"endDate": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			}
		},
		"securitySchemes": {
			"token": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "parking-api",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Jacek Złowocki"
		}
	},
	"paths": {
		"/users": {
			"get": {
				"operationId": "List",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/User"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [
					{
						"token": [
							"admin"
						]
					}
				],
				"parameters": []
			}
		},
		"/health": {
			"get": {
				"operationId": "Get",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/bookings": {
			"get": {
				"operationId": "List",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Booking"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [
					{
						"token": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Booking"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed"
					}
				},
				"security": [
					{
						"token": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/BookingPayload"
							}
						}
					}
				}
			}
		},
		"/bookings/{id}": {
			"get": {
				"operationId": "Get",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Booking"
								}
							}
						}
					}
				},
				"security": [
					{
						"token": []
					}
				],
				"parameters": []
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Booking"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed"
					}
				},
				"security": [
					{
						"token": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_BookingPayload_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"204": {
						"description": "No content"
					},
					"422": {
						"description": "Validation Failed"
					}
				},
				"security": [
					{
						"token": []
					}
				],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}